@@ -5,10 +5,10 @@
 
 <template>
 	<li class="question__item" @focusout="handleTabbing">
-		<div
+		<component
 			:is="pseudoIcon"
 			v-if="!isDropdown"
-			class="question__item__pseudoInput" />
+			class="question__item__pseudoInput"></component>
 		<input
 			ref="input"
 			:aria-label="ariaLabel"
@@ -149,13 +149,13 @@ export default {
 	},
 
 	emits: [
-		'tabbed-out',
-		'create-answer',
+		'tabbedOut',
+		'createAnswer',
 		'update:answer',
-		'focus-next',
+		'focusNext',
 		'delete',
-		'move-down',
-		'move-up',
+		'moveDown',
+		'moveUp',
 	],
 
 	data() {
@@ -250,7 +250,7 @@ export default {
 
 	methods: {
 		handleTabbing() {
-			this.$emit('tabbed-out', this.optionType)
+			this.$emit('tabbedOut', this.optionType)
 		},
 
 		/**
@@ -273,14 +273,15 @@ export default {
 
 				if (this.answer.local) {
 					// Dispatched for creation. Marked as synced
-					this.$set(this.answer, 'local', false)
+					// eslint-disable-next-line vue/no-mutating-props
+					this.answer.local = false
 					const newAnswer = await this.createAnswer(answer)
 
 					// Forward changes, but use current answer.text to avoid erasing
 					// any in-between changes while creating the answer
 					newAnswer.text = this.$refs.input.value
 
-					this.$emit('create-answer', this.index, newAnswer)
+					this.$emit('createAnswer', this.index, newAnswer)
 				} else {
 					await this.updateAnswer(answer)
 
@@ -300,7 +301,7 @@ export default {
 				return
 			}
 			if (this.index <= this.maxIndex) {
-				this.$emit('focus-next', this.index, this.optionType)
+				this.$emit('focusNext', this.index, this.optionType)
 			}
 		},
 
@@ -403,7 +404,7 @@ export default {
 		 * Reorder option but keep focus on the button
 		 */
 		onMoveDown() {
-			this.$emit('move-down')
+			this.$emit('moveDown')
 			this.focusButton(
 				this.index < this.maxIndex - 1
 					? 'buttonOptionDown'
@@ -412,7 +413,7 @@ export default {
 		},
 
 		onMoveUp() {
-			this.$emit('move-up')
+			this.$emit('moveUp')
 			this.focusButton(this.index > 1 ? 'buttonOptionUp' : 'buttonOptionDown')
 		},
 
