@@ -98,7 +98,7 @@ export default {
 		},
 	},
 
-	emits: ['share-form'],
+	emits: ['shareForm'],
 
 	setup() {
 		return {
@@ -186,7 +186,7 @@ export default {
 		},
 
 		onShareForm() {
-			this.$emit('share-form')
+			this.$emit('shareForm')
 		},
 	},
 }
